# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Tailoring Portfolio API
  description: Production-ready backend for tailoring portfolio with admin workspace and user wishlist
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-app.onrender.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/designs:
    get:
      summary: Get all designs (public)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of designs
...
  /api/admin/designs:
    post:
      summary: Create a new design (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, image]
              properties:
                title:
                  type: string
                description:
                  type: string
                tags:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Design created successfully
        '401':
          description: Unauthorized
        '403':
          description: Admin access required

    get:
      summary: Get all designs for admin management
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of designs
        '403':
          description: Admin access required
...
  /api/wishlist:
    get:
      summary: Get user's wishlist
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's wishlist
        '401':
          description: Unauthorized
...